bug: can't load models for deepseek beta Fixes #198
