+++ lib/data/res/openai.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:gpt_box/core/util/api_balance.dart';
 import 'package:gpt_box/data/model/chat/config.dart';
 import 'package:gpt_box/data/model/chat/github_model.dart';
+import 'package:gpt_box/data/res/l10n.dart';
 import 'package:gpt_box/data/res/url.dart';
 import 'package:gpt_box/data/store/all.dart';
 import 'package:openai_dart/openai_dart.dart';
@@ -83,6 +84,64 @@ abstract final class OpenAICfg {
     );
   }
 
+  /// Show the dialog to pick the model.
+  /// 
+  /// - [context] the context to show the dialog.
+  /// 
+  /// Return the model name if any model is picked, otherwise return null.
+  static Future<void> showPickModelDialog(BuildContext context) async {
+    if (OpenAICfg.current.key.isEmpty) {
+      context.showRoundDialog(
+        title: l10n.attention,
+        child: Text(l10n.needOpenAIKey),
+        actions: Btnx.oks,
+      );
+      return;
+    }
+
+    final model = await context.showPickSingleDialog(
+      items: OpenAICfg.models.value,
+      initial: OpenAICfg.current.model,
+      title: l10n.model,
+      actions: [
+        TextButton(
+          onPressed: () async {
+            context.pop();
+            await context.showLoadingDialog(
+                fn: () => OpenAICfg.updateModels(force: true),);
+            await showPickModelDialog(context);
+          },
+          child: Text(l10n.refresh),
+        ),
+        TextButton(
+          onPressed: () {
+            void onSave(String s) {
+              context.pop();
+              OpenAICfg.setTo(OpenAICfg.current.copyWith(model: s));
+            }
+
+            context.pop();
+            final ctrl = TextEditingController();
+            context.showRoundDialog(
+              title: l10n.custom,
+              child: Input(
+                controller: ctrl,
+                autoFocus: true,
+                onSubmitted: onSave,
+              ),
+              actions: Btn.ok(onTap: () => onSave(ctrl.text)).toList,
+            );
+          },
+          child: Text(l10n.custom),
+        ),
+      ],
+    );
+    if (model == null) return;
+    OpenAICfg.setTo(OpenAICfg.current.copyWith(model: model));
+    
+  }
+
+
   static void switchToDefault(BuildContext context) {
     final cfg = _store.fetch(ChatConfig.defaultId);
     if (cfg != null) return setTo(cfg);
@@ -126,8 +185,12 @@ abstract final class _ModelsCacher {
 
     final endpoint = OpenAICfg.current.url;
     // For most compatibility, use dio instead of openai_dart
+    final url = switch (endpoint) {
+      _ when endpoint.startsWith('https://api.deepseek.com/beta') => 'https://api.deepseek.com/v1/models',
+      _ => '$endpoint/models',
+    };
     final val = await myDio.get(
-      '$endpoint/models',
+      url,
       options: Options(
         headers: {'Authorization': 'Bearer ${OpenAICfg.current.key}'},
       ),
