+++ lib/view/page/home/appbar.dart
@@ -34,49 +34,46 @@ final class _CustomAppBar extends CustomAppBar {
       },
     );
 
-    final subtitle = ValBuilder(
-      listenable: OpenAICfg.vn,
-      builder: (val) {
-        return Text(
-          val.model.isEmpty ? libL10n.empty : val.model,
-          key: ValueKey(val),
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-          textAlign: TextAlign.left,
-          style: UIs.text12Grey,
-        );
-        // return ValBuilder(
-        //   listenable: _chatType,
-        //   builder: (type) {
-        //     return AnimatedSwitcher(
-        //       duration: _durationMedium,
-        //       switchInCurve: Easing.standardDecelerate,
-        //       switchOutCurve: Easing.standardDecelerate,
-        //       transitionBuilder: (child, animation) => SlideTransitionX(
-        //         position: animation,
-        //         child: FadeTransition(
-        //           opacity: animation,
-        //           child: child,
-        //         ),
-        //       ),
-        //       // Use a SizedBox to avoid the title jumping when switching chats.
-        //       child: Text(
-        //         switch (type) {
-        //           ChatType.text => val.model,
-        //           ChatType.img => val.imgModel,
-        //           //ChatType.audio => val.speechModel,
-        //         },
-        //         key: ValueKey(val),
-        //         maxLines: 1,
-        //         overflow: TextOverflow.ellipsis,
-        //         textAlign: TextAlign.left,
-        //         style: UIs.text12Grey,
-        //       ),
-        //     );
-        //   },
-        // );
-      },
-    );
+    final subtitle = OpenAICfg.vn.listenVal((val) {
+      return Text(
+        val.model.isEmpty ? libL10n.empty : val.model,
+        key: ValueKey(val),
+        maxLines: 1,
+        overflow: TextOverflow.ellipsis,
+        textAlign: TextAlign.left,
+        style: UIs.text12Grey,
+      );
+      // return ValBuilder(
+      //   listenable: _chatType,
+      //   builder: (type) {
+      //     return AnimatedSwitcher(
+      //       duration: _durationMedium,
+      //       switchInCurve: Easing.standardDecelerate,
+      //       switchOutCurve: Easing.standardDecelerate,
+      //       transitionBuilder: (child, animation) => SlideTransitionX(
+      //         position: animation,
+      //         child: FadeTransition(
+      //           opacity: animation,
+      //           child: child,
+      //         ),
+      //       ),
+      //       // Use a SizedBox to avoid the title jumping when switching chats.
+      //       child: Text(
+      //         switch (type) {
+      //           ChatType.text => val.model,
+      //           ChatType.img => val.imgModel,
+      //           //ChatType.audio => val.speechModel,
+      //         },
+      //         key: ValueKey(val),
+      //         maxLines: 1,
+      //         overflow: TextOverflow.ellipsis,
+      //         textAlign: TextAlign.left,
+      //         style: UIs.text12Grey,
+      //       ),
+      //     );
+      //   },
+      // );
+    });
 
     return CustomAppBar(
       centerTitle: false,
